[chore/#8] data/network 레이어 패키지 구조 변경
