+++ app/src/main/java/com/example/letssopt/LetsSoptApplication.kt
@@ -1,25 +1,21 @@
 package com.example.letssopt
 import android.app.Application
-import com.example.letssopt.data.local.AppDatabase
-import com.example.letssopt.data.network.RetrofitClient
-import com.example.letssopt.data.network.datasource.impl.AuthRemoteDataSourceImpl
-import com.example.letssopt.data.network.datasource.impl.UserRemoteDataSourceImpl
-import com.example.letssopt.data.repository.api.AuthRepository
-import com.example.letssopt.data.repository.api.StorageRepository
-import com.example.letssopt.data.repository.api.UserRepository
+import com.example.letssopt.core.network.RetrofitClient
+import com.example.letssopt.data.remote.datasource.impl.AuthRemoteDataSourceImpl
+import com.example.letssopt.data.remote.datasource.impl.UserRemoteDataSourceImpl
 
 class LetsSoptApplication : Application() {
     override fun onCreate() {
         super.onCreate()
         _root_ide_package_.com.example.letssopt.data.repository.api.AuthRepository.init(
             this,
-            _root_ide_package_.com.example.letssopt.data.network.datasource.impl.AuthRemoteDataSourceImpl(
-                _root_ide_package_.com.example.letssopt.data.network.RetrofitClient.authService
+            AuthRemoteDataSourceImpl(
+                RetrofitClient.authService
             )
         )
         _root_ide_package_.com.example.letssopt.data.repository.api.UserRepository.init(
-            _root_ide_package_.com.example.letssopt.data.network.datasource.impl.UserRemoteDataSourceImpl(
-                _root_ide_package_.com.example.letssopt.data.network.RetrofitClient.userService
+            UserRemoteDataSourceImpl(
+                RetrofitClient.userService
             )
         )
         _root_ide_package_.com.example.letssopt.data.repository.api.StorageRepository.init(
