+++ app/src/main/java/com/example/letssopt/data/remote/dto/GetUserProfileResponse.kt
@@ -1,10 +1,10 @@
-package com.example.letssopt.data.dto
+package com.example.letssopt.data.remote.dto
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class GetRecentUsersResponse(
+data class GetUserProfileResponse(
     @SerialName("success")
     val success: Boolean,
     @SerialName("status")
@@ -14,21 +14,21 @@ data class GetRecentUsersResponse(
     @SerialName("code")
     val code: String,
     @SerialName("data")
-    val data: RecentUsersDataList? = null,
+    val data: UserProfileData? = null,
 )
 
 @Serializable
-data class RecentUsersDataList(
-    @SerialName("users")
-    val users: List<RecentUsersData>
-)
-
-@Serializable
-data class RecentUsersData(
+data class UserProfileData(
     @SerialName("id")
     val id: Long,
+    @SerialName("loginId")
+    val logInId: String,
     @SerialName("name")
     val name: String,
+    @SerialName("email")
+    val email: String,
+    @SerialName("age")
+    val age: Int,
     @SerialName("part")
     val part: String,
 )
