+++ app/src/main/java/com/example/letssopt/data/remote/dto/PostSignInResponse.kt
@@ -1,9 +1,11 @@
-package com.example.letssopt.data.dto
+package com.example.letssopt.data.remote.dto
+
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
+
 @Serializable
-data class PostSignUpResponse(
+data class PostSignInResponse(
     @SerialName("success")
     val success: Boolean,
     @SerialName("status")
@@ -12,4 +14,12 @@ data class PostSignUpResponse(
     val message: String,
     @SerialName("code")
     val code: String,
+    @SerialName("data")
+    val data: UserData? = null,
+)
+
+@Serializable
+data class UserData(
+    @SerialName("userId")
+    val userId: Int,
 )
