+++ app/src/main/java/com/example/letssopt/data/remote/dto/PostSignUpResponse.kt
@@ -1,11 +1,9 @@
-package com.example.letssopt.data.dto
-
+package com.example.letssopt.data.remote.dto
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
-
 @Serializable
-data class PostSignInResponse(
+data class PostSignUpResponse(
     @SerialName("success")
     val success: Boolean,
     @SerialName("status")
@@ -14,12 +12,4 @@ data class PostSignInResponse(
     val message: String,
     @SerialName("code")
     val code: String,
-    @SerialName("data")
-    val data: UserData? = null,
-)
-
-@Serializable
-data class UserData(
-    @SerialName("userId")
-    val userId: Int,
 )
